backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
psycopg2-binary>=2.9,<3.0
django-cors-headers>=4.6,<5.0
gunicorn>=22.0,<23.0
google-auth>=2.29,<3.0
PyJWT[crypto]>=2.8,<3.0
requests>=2.31,<3.0
python-dotenv>=1.0,<2.0
openai>=1.0,<2.0
